requirements.txt
aerich==0.6.3
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.1
asyncpg==0.25.0
black==22.3.0
click==8.1.3
dictdiffer==0.9.0
fastapi==0.75.2
greenlet==2.0.0a2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
install==1.3.5
iso8601==1.0.2
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.0
pypika-tortoise==0.1.4
python-dotenv==0.20.0
pytz==2022.1
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
tomli==2.0.1
tomlkit==0.10.2
tortoise-orm==0.19.0
typing_extensions==4.2.0
uvicorn==0.17.6
zipp==3.8.0
